requirements.txt
pandas
confluent-kafka
psycopg2-binary
python-dotenv
pytest
Faker
-e .